+++ src/Filters/Filter.cpp
@@ -1,17 +1,17 @@
 #include "Filter.h"
 #include <algorithm>
 
-Vec4 Filter::getPixelAt(Vec4 *img, size_t x, size_t y, size_t width, size_t height)
+Vec4 Filter::getPixelAt(Vec4 *img, ssize_t x, ssize_t y, size_t width, size_t height)
 {
-	x = std::min(width - 1, std::max(size_t(0), x));
-	y = std::min(height - 1, std::max(size_t(0), y));
+	x = std::min(ssize_t(width - 1), std::max(ssize_t(0), x));
+	y = std::min(ssize_t(height - 1), std::max(ssize_t(0), y));
 	return img[x + y * width];
 }
 
-float Filter::getZAt(float *zBuffer, size_t x, size_t y, size_t width, size_t height)
+float Filter::getZAt(float *zBuffer, ssize_t x, ssize_t y, size_t width, size_t height)
 {
-	x = std::min(width - 1, std::max(size_t(0), x));
-	y = std::min(height - 1, std::max(size_t(0), y));
+	x = std::min(ssize_t(width - 1), std::max(ssize_t(0), x));
+	y = std::min(ssize_t(height - 1), std::max(ssize_t(0), y));
 	return zBuffer[x + y * width];
 }
